Renames and file layout from ringbahn
