+++ src/cqe.rs
@@ -1,114 +1,31 @@
 use std::io;
-use std::marker::PhantomData;
-use std::mem::MaybeUninit;
-use std::ptr::{self, NonNull};
+use std::ptr::NonNull;
 
-use super::{IoUring, resultify};
-
-/// The queue of completed IO events.
-///
-/// Each element is a [`CompletionQueueEvent`](crate::cqe::CompletionQueueEvent).
-///
-/// Completion does not imply success. Completed events may be [timeouts](crate::cqe::CompletionQueueEvent::is_timeout).
-pub struct CompletionQueue<'ring> {
-    ring: NonNull<uring_sys::io_uring>,
-    _marker: PhantomData<&'ring mut IoUring>,
-}
-
-impl<'ring> CompletionQueue<'ring> {
-    pub(crate) fn new(ring: &'ring IoUring) -> CompletionQueue<'ring> {
-        CompletionQueue {
-            ring: NonNull::from(&ring.ring),
-            _marker: PhantomData,
-        }
-    }
-
-    pub fn peek_for_cqe(&mut self) -> Option<CompletionQueueEvent> {
-        unsafe {
-            let mut cqe = MaybeUninit::uninit();
-            let count = uring_sys::io_uring_peek_batch_cqe(self.ring.as_ptr(), cqe.as_mut_ptr(), 1);
-            if count > 0 {
-                Some(CompletionQueueEvent::new(self.ring, &mut *cqe.assume_init()))
-            } else {
-                None
-            }
-        }
-    }
-
-    pub fn wait_for_cqe(&mut self) -> io::Result<CompletionQueueEvent> {
-        self.wait_for_cqes(1)
-    }
-
-    pub fn wait_for_cqes(&mut self, count: u32) -> io::Result<CompletionQueueEvent> {
-        unsafe {
-            let mut cqe = MaybeUninit::uninit();
-
-            resultify(uring_sys::io_uring_wait_cqes(
-                self.ring.as_ptr(),
-                cqe.as_mut_ptr(),
-                count as _,
-                ptr::null(),
-                ptr::null(),
-            ))?;
-
-            Ok(CompletionQueueEvent::new(self.ring, &mut *cqe.assume_init()))
-        }
-    }
-
-    pub fn cqes(&mut self) -> CompletionQueueEvents<'ring, '_> {
-        CompletionQueueEvents {
-            queue: self,
-            ready: 0,
-        }
-    }
-
-    pub fn cqes_blocking(&mut self) -> CompletionQueueEventsBlocking<'ring, '_> {
-        CompletionQueueEventsBlocking {
-            queue: self,
-            ready: 0,
-        }
-    }
-
-    pub fn ready(&self) -> u32 {
-        unsafe { uring_sys::io_uring_cq_ready(self.ring.as_ptr()) }
-    }
-
-    pub fn eventfd_enabled(&self) -> bool {
-        unsafe { uring_sys::io_uring_cq_eventfd_enabled(self.ring.as_ptr()) }
-    }
-
-    pub fn eventfd_toggle(&mut self, enabled: bool) -> io::Result<()> {
-        resultify(unsafe { uring_sys::io_uring_cq_eventfd_toggle(self.ring.as_ptr(), enabled) })?;
-        Ok(())
-    }
-}
-
-unsafe impl<'ring> Send for CompletionQueue<'ring> { }
-unsafe impl<'ring> Sync for CompletionQueue<'ring> { }
+use super::{CompletionQueue, resultify};
 
 /// A completed IO event.
-pub struct CompletionQueueEvent {
+pub struct CQE {
     user_data: u64,
     res: i32,
     flags: u32,
 }
 
-impl CompletionQueueEvent {
-    pub fn from_raw(cqe: uring_sys::io_uring_cqe) -> CompletionQueueEvent {
-        CompletionQueueEvent {
+impl CQE {
+    pub fn from_raw(cqe: uring_sys::io_uring_cqe) -> CQE {
+        CQE {
             user_data: cqe.user_data,
             res: cqe.res,
             flags: cqe.flags,
         }
     }
 
-    pub fn from_raw_parts(user_data: u64, res: i32, flags: u32) -> CompletionQueueEvent {
-        CompletionQueueEvent {
+    pub fn from_raw_parts(user_data: u64, res: i32, flags: u32) -> CQE {
+        CQE {
             user_data, res, flags,
         }
     }
 
-    pub(crate) fn new(ring: NonNull<uring_sys::io_uring>, cqe: &mut uring_sys::io_uring_cqe) -> CompletionQueueEvent {
+    pub(crate) fn new(ring: NonNull<uring_sys::io_uring>, cqe: &mut uring_sys::io_uring_cqe) -> CQE {
         let user_data = cqe.user_data;
         let res = cqe.res;
         let flags = cqe.flags;
@@ -117,12 +34,12 @@ impl CompletionQueueEvent {
             uring_sys::io_uring_cqe_seen(ring.as_ptr(), cqe);
         }
 
-        CompletionQueueEvent::from_raw_parts(user_data, res, flags)
+        CQE::from_raw_parts(user_data, res, flags)
     }
 
     /// Check whether this event is a timeout generated by this crate.
     /// ```
-    /// # use iou::{IoUring, SubmissionQueueEvent, CompletionQueueEvent};
+    /// # use iou::{IoUring, SubmissionQueueEvent, CQE};
     /// # fn main() -> std::io::Result<()> {
     /// # let mut ring = IoUring::new(2)?;
     /// # let mut sqe = ring.next_sqe().unwrap();
@@ -134,7 +51,7 @@ impl CompletionQueueEvent {
     /// # let mut cq_event;
     /// cq_event = ring.wait_for_cqe()?;
     /// # // rewrite to be a fake timeout
-    /// # let cq_event = CompletionQueueEvent::from_raw_parts(uring_sys::LIBURING_UDATA_TIMEOUT, 0, 0);
+    /// # let cq_event = CQE::from_raw_parts(uring_sys::LIBURING_UDATA_TIMEOUT, 0, 0);
     /// assert!(cq_event.is_iou_timeout());
     /// # Ok(())
     /// # }
@@ -160,16 +77,16 @@ impl CompletionQueueEvent {
     }
 }
 
-unsafe impl Send for CompletionQueueEvent { }
-unsafe impl Sync for CompletionQueueEvent { }
+unsafe impl Send for CQE { }
+unsafe impl Sync for CQE { }
 
-pub struct CompletionQueueEvents<'a, 'b> {
-    queue: &'b mut CompletionQueue<'a>,
-    ready: u32,
+pub struct CQEs<'a, 'b> {
+    pub(crate) queue: &'b mut CompletionQueue<'a>,
+    pub(crate) ready: u32,
 }
 
-impl Iterator for CompletionQueueEvents<'_, '_> {
-    type Item = CompletionQueueEvent;
+impl Iterator for CQEs<'_, '_> {
+    type Item = CQE;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.ready == 0 {
@@ -184,13 +101,13 @@ impl Iterator for CompletionQueueEvents<'_, '_> {
     }
 }
 
-pub struct CompletionQueueEventsBlocking<'a, 'b> {
-    queue: &'b mut CompletionQueue<'a>,
-    ready: u32,
+pub struct CQEsBlocking<'a, 'b> {
+    pub(crate) queue: &'b mut CompletionQueue<'a>,
+    pub(crate) ready: u32,
 }
 
-impl Iterator for CompletionQueueEventsBlocking<'_, '_> {
-    type Item = io::Result<CompletionQueueEvent>;
+impl Iterator for CQEsBlocking<'_, '_> {
+    type Item = io::Result<CQE>;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.ready == 0 {
